% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pgNbrs.R
\name{collapse_pgNbrs}
\alias{collapse_pgNbrs}
\title{The program collapse page numbers that belong to the same family (wrt Dataset and Variable columns).}
\usage{
collapse_pgNbrs(dsin = NULL, domain_list = NULL)
}
\arguments{
\item{dsin}{The preprocessed dataset with all page numbers for each domain extracted from aCRF file.}

\item{domain_list}{List of SDTM domains read from aCRF.}
}
\description{
The program collapse page numbers that belong to the same family (wrt Dataset and Variable columns).
}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.worku@ocs-consulting.com>
}
