% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sdtmVars_pgNbr.R
\name{get_sdtmVars_pgNbr}
\alias{get_sdtmVars_pgNbr}
\title{This is the main program which does the page extraction from aCRF page.}
\usage{
get_sdtmVars_pgNbr(crf_pageIn = NULL, defineOrigin_variableTab = NULL,
  domain_list = NULL, until_pgNbr = NULL)
}
\arguments{
\item{crf_pageIn}{The preprocessed dataset with all text read from aCRF using pkg::pdftools.}

\item{defineOrigin_variableTab}{The preprocessed dataset with define origin info.}

\item{domain_list}{List of SDTM domains read from aCRF.}

\item{until_pgNbr}{Included for debugging purpose.}
}
\description{
This is the main program which does the page extraction from aCRF page.
}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.worku@ocs-consulting.com>
}
