% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valueLevel_pgNbr.R
\name{get_valueLevel_pgNbr}
\alias{get_valueLevel_pgNbr}
\title{This is the main program which does the page extraction from aCRF page, related to Value level tab.}
\usage{
get_valueLevel_pgNbr(crf_pageIn = NULL, defineOrigin_valueLevelTab = NULL,
  until_pgNbr = NULL)
}
\arguments{
\item{crf_pageIn}{The preprocessed dataset with all text read from aCRF using pkg::pdftools.}

\item{until_pgNbr}{Included for debugging purpose.}

\item{defineOrigin_variableTab}{The preprocessed dataset with define specs info.}
}
\description{
This is the main program which does the page extraction from aCRF page, related to Value level tab.
}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.worku@ocs-consulting.com>
}
