% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_crf_fn.R
\name{preprocess_crf}
\alias{preprocess_crf}
\title{The program splits raw data read from aCRF file using page delimiter depending which operating system is the package is running from. 
It can handle both Windows and Mac systems.}
\usage{
preprocess_crf(crf_txt_raw = NULL)
}
\arguments{
\item{crf_txt_raw}{The raw info read from aCRF file.}
}
\description{
The program splits raw data read from aCRF file using page delimiter depending which operating system is the package is running from. 
It can handle both Windows and Mac systems.
}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.worku@ocs-consulting.com>
}
