% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCRFPageMinner_main.R
\name{run_aCRFPageMinner}
\alias{run_aCRFPageMinner}
\title{Extracts page numbers of SDTM variables available in manually created aCRF file. The output is used for
creation of Define XML file.}
\usage{
run_aCRFPageMinner()
}
\description{
The input files (e.g., studyid, aCRF file, and Define Origin file) are taken from a user via console.
}
\examples{
## library(aCRFPageMinner)
## run_aCRFPageMinner()

}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.worku@ocs-consulting.com>
}
